requirements.txt
requests
python-dotenv
numpy
pillow
scikit-learn
matplotlib
pandas